Implement malloc() functionality for buffer

+++ cbuf_int.c
@@ -88,17 +88,31 @@ cbuf_t *cbuf_create_(uint32_t obj_sz,
 
 	/* MMAP */
 	// TODO Robert: implement malloc case
+	
 	char tfile[] = "/tmp/cbufXXXXXX";
-	b->mmap_fd = mkostemp(tfile, O_NOATIME);
-	Z_die_if(b->mmap_fd < 1, "mmap '%s'", tfile);
-	/* make space, map */
-	size_t len = next_multiple(buf_sz, cbuf_hugepage_sz);
-	Z_die_if(ftruncate(b->mmap_fd, len), "");
-	/* MUST be MAP_SHARED. If not, cbuf -> file splices WILL NOT WRITE to disk. */
-	b->buf = mmap(NULL, len, (PROT_READ | PROT_WRITE), 
-		(MAP_SHARED | MAP_LOCKED | MAP_NORESERVE), b->mmap_fd, 0);
-	Z_die_if(b->buf == MAP_FAILED, "sz:%ld", len);
-	Z_die_if(unlink(tfile), "");
+	if (b->cbuf_flags & CBUF_MALLOC && !(b->cbuf_flags & CBUF_P)){
+		Z_inf(3, "Just a print for running a malloc'ed case");
+		size_t len = next_multiple(buf_sz, cbuf_hugepage_sz);
+		/* This is a malloc'ed cbuf. We will read and write to the cbuf
+		 * and then vmsplice to the next pipe for sending over the "wire". */
+		/*  I changed this to an int as the cbuf holds pointers not characters in this case. */
+		b->buf = (char *)malloc(len);
+		/*   char tfile[] = "/tmp/cbufXXXXXX"; */
+		Z_die_if((b->buf == NULL), ""); 
+	} else {		
+		Z_die_if(b->cbuf_flags & CBUF_MALLOC, "malloc not implemented");
+		/*   char tfile[] = "/tmp/cbufXXXXXX"; */
+		b->mmap_fd = mkostemp(tfile, O_NOATIME);
+		Z_die_if(b->mmap_fd < 1, "mmap '%s'", tfile);
+		/* make space, map */
+		size_t len = next_multiple(buf_sz, cbuf_hugepage_sz);
+		Z_die_if(ftruncate(b->mmap_fd, len), "");
+		/* MUST be MAP_SHARED. If not, cbuf -> file splices WILL NOT WRITE to disk. */
+		b->buf = mmap(NULL, len, (PROT_READ | PROT_WRITE), 
+			(MAP_SHARED | MAP_LOCKED | MAP_NORESERVE), b->mmap_fd, 0);
+		Z_die_if(b->buf == MAP_FAILED, "sz:%ld", len);
+		Z_die_if(unlink(tfile), "");
+ 	}
 
 	Z_inf(3, "cbuf @0x%lx size=%d obj_sz=%d overflow_=0x%x sz_bitshift_=%d", 
 	     (uint64_t)b, cbuf_sz_buf(b), cbuf_sz_obj(b), b->overflow_, b->sz_bitshift_);
@@ -126,23 +140,29 @@ void cbuf_free_(cbuf_t *buf)
 	}
 
 	/* free memory */
-	// TODO Robert: handle mmap case
+	// TODO Robert: handle malloc() case
 	if (buf->buf) {
+		/*  handle the malloc'ed case */
 
-		/* handle backing store (cbufp_) ? */
-		if (buf->cbuf_flags & CBUF_P) {
+		Z_inf(3, "Just a print for running a malloc'ed case");
+		if (!(buf->cbuf_flags & CBUF_MALLOC && !(buf->cbuf_flags & CBUF_P))){
+			/* handle backing store (cbufp_) ? */
+			if (buf->cbuf_flags & CBUF_P) {
 			cbufp_t *f = buf->buf;
 			sbfu_unmap(f->fd, &f->iov);
 			unlink(f->file_path); 
 			free(f->file_path);
 			errno = 0;
+			}
+
+			/* avoid trying to free a '-1' (aka: MAP_FAILED) */
+			if (buf->buf != MAP_FAILED)
+				munmap(buf->buf, next_multiple(cbuf_sz_buf(buf), cbuf_hugepage_sz));
 		}
-
-		/* avoid trying to free a '-1' (aka: MAP_FAILED) */
-		if (buf->buf != MAP_FAILED)
-			munmap(buf->buf, next_multiple(cbuf_sz_buf(buf), cbuf_hugepage_sz));
+		else	
+			Z_inf(3, "Just a print for running a malloc'ed case");
 	}
-
+	
 	/* open file descriptors */
 	if (buf->mmap_fd)
 		close(buf->mmap_fd);
